requirements.txt
pandas
openai
psycopg2
python-dotenv
timescale-vector
instructor
anthropic
transformers
torch
